+++ my-app/client/src/pages/PinPage/PinPage.tsx
