+++ my-app/client/src/pages/ScavengerPage/scavengeScreen.css
