+++ my-app/client/src/player-utils.ts
@@ -1,5 +1,5 @@
 
-import { Player, Item } from "./types";
+import { Player, Item } from "./types/types";
 import axios from 'axios';
 export const API_BASE_URL = "http://localhost:5000";
